+++ CHANGELOG.md
@@ -10,6 +10,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Export genotype with gene variantS search (#6088)
 - igv.js version to 3.7.3 (#6096)
+- Refactored `scout delete` commands and relative tests into separate modules (#6102)
 ### Fixed
 - Comments' text wrapping in ACMG classifications exported as PDF (#6086)